requirements.txt
jupyter
pymongo
python-dotenv
